Resolve on NeoForge https://github.com/pcal43/fastback/issues/260 ed25519 keys are not support in fasback-forge
At-least I believe it does. Uploading was failing with a jgit error aboud failing to negotiate the algorithm.
Uploading to gitea and github now work after this change, which indicates to me that this issue is resolved.

All that's been changed is neoforge dependencies have been updated to reflect the dependencies in the fabric version.