Resolve NeoForge equivalent of 'https://github.com/pcal43/fastback/issues/260 ed25519 keys are not support in fasback-forge'
At-least I believe it does. Uploading was failing with a jgit error aboud failing to negotiate the algorithm.
Uploading to gitea and github now work after this change, which indicates to me that this issue is resolved.

However, org.apache.sshd goes missing in development environments, and the SshHacks fix does not work anymore because of the library change.
I've made an issue in architectury-loom for this: https://github.com/architectury/architectury-loom/issues/248

In addition, the error thrown due to org.apache.sshd going missing is now being logged, but will not prevent fastback from loading.
Allowing the development environment to still be used without support for ssh.